Fix Rosetta SQL parsing int4 overflow issue (#4070)
- Add explicit ::bigint conversion in named SQL param arithmetic operation

+++ hedera-mirror-rosetta/app/persistence/block.go
@@ -81,7 +81,7 @@ const (
                                              coalesce((
                                                select consensus_start-1
                                                from record_file
-                                               where index = @index + 1
+                                               where index = @index + 1::bigint
                                              ), consensus_end) as consensus_end,
                                              hash,
                                              index,
